requirements.txt
fastapi
uvicorn
aiohttp
redis
discord.py
numpy
pandas
scikit-learn